Feat/element init

+++ src/main/java/com/leeroy/forwordpanel/forwordpanel/controller/ServerController.java
@@ -1,6 +1,8 @@
 package com.leeroy.forwordpanel.forwordpanel.controller;
 
+import com.leeroy.forwordpanel.forwordpanel.common.WebCurrentData;
 import com.leeroy.forwordpanel.forwordpanel.common.response.ApiResponse;
+import com.leeroy.forwordpanel.forwordpanel.dto.PageRequest;
 import com.leeroy.forwordpanel.forwordpanel.model.Port;
 import com.leeroy.forwordpanel.forwordpanel.model.Server;
 import com.leeroy.forwordpanel.forwordpanel.service.PortService;
@@ -16,21 +18,11 @@ public class ServerController {
     @Autowired
     private ServerService serverService;
 
-    /**
-     * 页面
-     *
-     * @return
-     */
-    @RequestMapping("/manage")
-    public String portManage() {
-        return "server/index";
-    }
-
 
     @ResponseBody
-    @GetMapping("getPage")
-    public ApiResponse getPage() {
-        return ApiResponse.ok(serverService.findList());
+    @PostMapping("getPage")
+    public ApiResponse getPage(@RequestBody PageRequest pageRequest) {
+        return ApiResponse.ok(serverService.getServerPage(pageRequest));
     }
 
 
@@ -43,14 +35,19 @@ public class ServerController {
     @ResponseBody
     @RequestMapping(value = "save", method = RequestMethod.POST)
     public ApiResponse saveUserPort(@RequestBody Server server) {
+        if(WebCurrentData.getUser().getUserType()>0){
+            return ApiResponse.error("403", "您没有权限执行此操作");
+        }
         return serverService.save(server);
     }
 
     @ResponseBody
-    @PostMapping("delete")
+    @GetMapping("delete")
     public ApiResponse delete(Integer id) {
-        serverService.delete(id);
-        return ApiResponse.ok();
+        if(WebCurrentData.getUser().getUserType()>0){
+            return ApiResponse.error("403", "您没有权限执行此操作");
+        }
+        return serverService.delete(id);
     }
